restructure types into its own directory and separate modules
